kernel: sched: idle: fix possible alignment fault

+++ kernel/sched/idle.c
@@ -209,7 +209,7 @@ DEFINE_PER_CPU(bool, cpu_dead_idle);
  */
 static void cpu_idle_loop(void)
 {
-	unsigned int cpu = smp_processor_id();
+    uintptr_t cpu = smp_processor_id();
 
 	while (1) {
 		/*
